Restructured TagCloud
->Geometry moved to TagCloud root and renamed to GeometryUtils
-> Algorithms renamed to LayoutAlgorithms
-> interface iLayoter moved to LayoutAlgorithms

first attempt to adapt visualizer to drawing words.
