+++ TagCloudContainerTask/TagCloud/LayoutAlgorithms/AlgorithmFromTDD/Layouting/CircularLayouter.cs
@@ -2,9 +2,9 @@
 using System.Collections.Generic;
 using System.Drawing;
 using System.Linq;
-using TagCloud.Algorithms.AlgorithmFromTDD.Geometry;
+using TagCloud.GeometryUtils;
 
-namespace TagCloud.Algorithms.AlgorithmFromTDD.Layouting
+namespace TagCloud.LayoutAlgorithms.AlgorithmFromTDD.Layouting
 {
     public class CircularLayouter : ICloudLayouter
     {
