+++ TagCloudContainerTask/TagCloud/LayoutAlgorithms/AlgorithmFromTDD/Layouting/ICloudLayouter.cs
