+++ TagCloudContainerTask/TagCloud/TagCloud.csproj
@@ -43,12 +43,12 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Algorithms\AlgorithmFromTDD\Geometry\DirectingArrow.cs" />
-    <Compile Include="Algorithms\AlgorithmFromTDD\Geometry\PointExtension.cs" />
-    <Compile Include="Algorithms\AlgorithmFromTDD\Geometry\RectangleExtension.cs" />
-    <Compile Include="Algorithms\AlgorithmFromTDD\Layouting\CircularLayouter.cs" />
-    <Compile Include="Algorithms\AlgorithmFromTDD\Layouting\ICloudLayouter.cs" />
-    <Compile Include="Algorithms\AlgorithmFromTDD\Layouting\ILayouter.cs" />
+    <Compile Include="GeometryUtils\DirectingArrow.cs" />
+    <Compile Include="GeometryUtils\PointExtension.cs" />
+    <Compile Include="GeometryUtils\RectangleExtension.cs" />
+    <Compile Include="LayoutAlgorithms\AlgorithmFromTDD\Layouting\CircularLayouter.cs" />
+    <Compile Include="LayoutAlgorithms\AlgorithmFromTDD\Layouting\ICloudLayouter.cs" />
+    <Compile Include="LayoutAlgorithms\ILayouter.cs" />
     <Compile Include="Saving\BitmapSaver.cs" />
     <Compile Include="Saving\IBitmapSaver.cs" />
     <Compile Include="Visualization\Drawer.cs" />
